works with inkscape limitations
only invert.py was worth something, but inkscape can't do shit without the gui so i removed it (and all the other crap)
invert.py called inkscape to invert everything at once like this:

    # Invert the entire SVG
    inverted_svg = tmp_dir / f"{basename}_inverted.svg"
    invert_command = (
        f"inkscape {svg_file} "
        f"--batch-process "
        f"--actions=\"select-all;org.inkscape.effect.filter.Invert.noprefs;export-filename:{inverted_svg};export-do;\" "
        f"--pdf-font-strategy=substitute"
    )

The problem now is that inkscape can't handle all pdfs. Some images for example are lost. Ill make an issue with that

seing over here what other tools there are to make the job:
https://gist.github.com/douglasmiranda/9c19f23c4570a7b7e02137791880ab43

mupdf also loses those weird images that inkscape loses as well
and pdf2svg converts everything to raster, thats exactly what I didnt want

for now the best tool for the job seem to be online convertrs: https://www.pdfconvertonline.com/invert-pdf/

Also fonts, using --pdf-font-strategy=substitute, so it might not find some fonts and it might look a bit weird, I'd have to see a way to try to add a lot of fonts to prevent that or to download on demand tho that sounds hard af